+++ src/App.jsx
@@ -9,7 +9,7 @@ import ConfirmationModal from './components/modals/ConfirmationModal'
 import { useMotorSync } from './hooks/useMotorSync'
 
 // App version
-const APP_VERSION = '0.2.5-dev'
+const APP_VERSION = '0.2.6-dev'
 
 function App() {
   const {